+++ FTMobileSDK/FTSDKCore/FTRUM/RUMCore/FTRUMManager.h
@@ -101,25 +101,27 @@ typedef NS_ENUM(NSUInteger, FTLaunchType) {
  */
 -(void)stopViewWithProperty:(nullable NSDictionary *)property;
 
-#pragma mark - action -
-/// 点击事件
-/// @param actionName actionName 点击的事件名称
-- (void)addClickActionWithName:(nonnull NSString *)actionName;
-/**
- * 点击事件
- * @param actionName 点击的事件名称
- */
-- (void)addClickActionWithName:(NSString *)actionName property:(nullable NSDictionary *)property;
-/// action 事件
-/// @param actionName 事件名称
-/// @param actionType 事件类型
-- (void)addActionName:(nonnull NSString *)actionName actionType:(nonnull NSString *)actionType;
+#pragma mark - Action -
 
-/// action 事件
-/// @param actionName 事件名称
-/// @param actionType 事件类型
-/// @param property 事件自定义属性(可选)
-- (void)addActionName:(NSString *)actionName actionType:(NSString *)actionType property:(nullable NSDictionary *)property;
+/// 启动 RUM Action。
+///
+/// RUM 会绑定该 Action 可能触发的 Resource、Error、LongTask 事件。避免在 0.1 s 内多次添加，同一个 View 在同一时间只会关联一个 Action，在上一个 Action 未结束时，新增的 Action 会被丢弃。
+/// 与 `addAction:actionType:property` 方法添加 Action 互不影响。
+///
+/// - Parameters:
+///   - actionName: 事件名称
+///   - actionType: 事件类型
+///   - property: 事件自定义属性(可选)
+- (void)startAction:(NSString *)actionName actionType:(NSString *)actionType property:(nullable NSDictionary *)property;
+
+/// 添加 Action 事件.无 duration，无丢弃逻辑
+///
+/// 与 `startAction:actionType:property:` 启动的 RUM Action 互不影响。
+/// - Parameters:
+///   - actionName: 事件名称
+///   - actionType: 事件类型
+///   - property: 事件自定义属性(可选)
+- (void)addAction:(NSString *)actionName actionType:(NSString *)actionType property:(nullable NSDictionary *)property;
 /**
  * 应用启动
  * @param type      启动类型